increament chatroom send message function

+++ eCommunicationBook/Helper/DateHelper.swift
@@ -8,11 +8,11 @@
 import Foundation
 
 extension Date {
-    var millisecondsSince1970: Int64 {
-        return Int64((self.timeIntervalSince1970 * 1000.0).rounded())
+    var millisecondsSince1970: Double {
+        return (self.timeIntervalSince1970 * 1000.0).rounded()
     }
     
-    init(milliseconds: Int64) {
+    init(milliseconds: Double) {
         self = Date(timeIntervalSince1970: TimeInterval(milliseconds / 1000))
     }
